Lots and lots of stuff.  Need to stub out more of the character creation process and at least get the character list working.

+++ interfaces/worldobject.go
@@ -4,5 +4,5 @@ import "github.com/tobz/phosphorus/utils"
 
 type WorldObject interface {
 	Position() utils.Point3D
-	ObjectID() uint32
+	ObjectID() uint16
 }
